Move chat unrelated menu items to Sidebar
