+++ scss/_sidebar.scss
@@ -29,6 +29,18 @@
     animation: sidebar-slide-out 0.5s forwards;
     -webkit-animation: sidebar-slide-out 0.5s forwards;
   }
+
+  .sidebar-item {
+    height: 40px;
+    width: 100%;
+    /* background-color: lightgrey; */
+    font-size: 20px;
+    line-height: 40px;
+    padding-left: 30px;
+    &:hover {
+      background-color: lightgrey
+    }
+  }
 }
 
 @keyframes sidebar-slide-in {
